backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.6.0
django-celery-results==2.5.1

psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0

gunicorn==22.0.0
whitenoise==6.6.0

reportlab==4.2.0
Pillow==10.3.0

drf-spectacular==0.27.2

python-decouple==3.8
dj-database-url==2.1.0
